src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
aniso8601==10.0.0
    # via flask-restx
anyio==4.9.0
    # via httpx
attrs==25.1.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via flask
cachelib==0.13.0
    # via flask-session
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
click==8.1.8
    # via flask
dependency-injector==4.46.0
    # via -r requirements.in
factory-boy==3.3.3
    # via -r requirements.in
faker==37.1.0
    # via
    #   -r requirements.in
    #   factory-boy
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-marshmallow
    #   flask-restx
    #   flask-session
flask-marshmallow==1.3.0
    # via -r requirements.in
flask-restx==1.3.0
    # via -r requirements.in
flask-session==0.8.0
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
importlib-resources==6.5.2
    # via flask-restx
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jsonschema==4.23.0
    # via flask-restx
jsonschema-specifications==2024.10.1
    # via jsonschema
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.26.0
    # via
    #   flask-marshmallow
    #   marshmallow-sqlalchemy
marshmallow-sqlalchemy==1.4.2
    # via -r requirements.in
msgspec==0.19.0
    # via flask-session
packaging==24.2
    # via marshmallow
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2024.2
    # via flask-restx
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.40
    # via marshmallow-sqlalchemy
typing-extensions==4.12.2
    # via sqlalchemy
tzdata==2025.2
    # via faker
werkzeug==3.1.3
    # via
    #   flask
    #   flask-restx
